Added dynamic program context via JSON files
This commit introduces the ability to craft your own program context
using JSON files. The JSON file contains an array of contexts which are
built up from memory blocks, structs, pointers and itegers.

Crafting these manually can be a lot of work, hopefully we can automate
a large part of this with tooling to capture and convert actual data to
our context file format.