Added breakpoints and continue commands
We now have breakpoints, which can work both on line and instruction
level. The `continue` command is used to run a program until a
breakpoint is hit or it exits.

Used the locspec format defined by delve(and their library). We don't
support all of the formats yet. Also missing a few ease of use features
which will be added in a later commit