Added macro command and comments
Added the `macro` debugger command. Macros can be used as you would
expect, to execute a series of commands. Macros can be recorded with the
`macro start` command, any commands executed will be recorded. Once the
`macro stop` command is called, a macro is created. A list of macros
loaded macros can be displayed with the `macro list` command and the
contents of a macro displayed with the `macro show` command. The
commands in a loaded macro can be exectued/replayed with the `macro exec`
command.

Macros can be persisted to a file with the `macro save` command and
loaded into a session with the `macro load` command. A file can contain
muliple macros and comments. The executor now ignores lines starting
with # or // but will still record the into a macro, allowing a user to
document a macro while recording. When playing back the comments are
also displayed back.

A file can have multiple macros, when saving we will rewrite a whole
config file but only change the macro we are saving or appending it to
the file. Comments in macro files are preserved. If a file already
exists on disk, we check for the existance of a magic comment at the
start of the file, this is to avoid accidental corruption of non-macro
files. Files without the header can still be loaded, just not saved to.

The `macro run` command can be used to execute all macros in a given
file without having to load them and then execute, saving keystrokes,
which is desirable for macros to quickly setup the debugger for a
specific program.

The `macro set` and `macro del` commands can be used to edit a macro
from within debugger, which is useful for macros recorded to be used in
the same session without saving them. Typos can be corrected or invalid
commands deleted without having to rerecord the macro.

Lastly, the `macro un-load` command can be used to unload command, in
case one might want to rerecord a macro with the same name(which is now
allowed otherwise)