Added source file annotation to the `list-instructions` output
The `list-instructions` output is now annotated with the lines from the
source code to which a block of instructions belong. The formatting is
inspired by the output format of `llvm-objdump -S <ELF file>`.

This feature is especially useful for larger programs where it might
harder to keep track of what instructions are actually doing. The
alternative was executing both `list` and `list-instructions`, this is
more elagant with little downside.