Added BTF map k/v information to format the `map read-all` output
The actual data of map keys and values are just byte slices, the meaning
of these bytes are known to the program, but not directly visable to a
debugger. We can use the BTF type information to see what the key and
value types are and how to decode them into human readable
reprentation(structs, ints, floats).

If no BTF info is available we fall back to just showing the unformatted
bytes.