Added `map get` and `map del` commands
The `map get` command takes a map id/name and a key and outputs the
value if any is found. The output of the get command is pretty printed
which makes it easier to read the BTF decoded value if it is large
and/or deeply nested.

The `map del` command takes a map id/name and a key and will delete the
given key in the map.