Made ELF loading consistant between debugging sessions
Since the ELF maps and programs are stored in maps, the order of k/v
pairs is not garanteed. This meant that each run, the programs and maps
could get different ids since they are dependant on the order they are
loaded in. We now always load programs and maps in ascending order based
on their names. This allows macros to use indexes which before was
impossible due to the changing meaning of them.