Replaced gobpfld with cilium/ebpf and mimic
This commit replaces the gobpfld library with the cilium/ebpf library
which is much better maintained. cilium/ebpf has no emulator, it is out
of scope for the project. So I took the gobpfld/emulator package, and
rewrote it based on cilium/ebpf, this new emulator is christened 'Mimic'
and lives in its own repository.

I was able to preserve most functionality, except the `locals` command
which wan't very useful anyway. The most major difference is he way
program instructions are formatted, which now used the formatting from
cilium/ebpf. In the future I will likely make the formatting of
instruction configurable so we can support both cilium/ebpf style and
llvm style instructions.

Another noticable change is to the context format, which is now defined
in Mimic instread of in the debugger.