WIP program instrumentation for context capture
This is a non working in progress commit for save keeping. The idea
behind this commit is to load a ELF program into the kernel and pin it
to the BPFFS. Before loading, we modify the program to capture the
inputs and outputs of helper functions as well as the given context on
startup, basically everything interaction with the linux kernel. This
instrumentation code will then send this over a PERF_EVENT_BUFFER map.

The proposed workflow being:
1. Load ELF program using edb
2. Start edb context capture command
3. Start normal loader program (which should be instructed to use a
   pinned program and maps instread of parsing the ELF itself)
4. Capture a senario one would like to debug/inspect
5. Stop the edb context capture command (and loader), which will flush
   any pending data to a context file.
6. Start a edb session with the ELF file and the gathered context file.

Since programs will always behave the same given the same inputs, we
should be able to emulate what happens in the kernel one-to-one.