Added initial program analyse functions and graph command
Program analysis is requried for the capture context feature. The
current iteration uses hardcoded stack locations to store temporary
variables. We do this to avoid conflicts with the existing program,
assuming they will never use fp-256. However, by doing so, we force the
program to allocate up to 256 bytes of stack for each frame, causing
instrumented programs to exceed the max stack size when using bpf-to-bpf
functions.

The plan to solve this issue is to replicate part of the static-analysis
the kernel does, specifically tracking which registers and stack
locations will be used by a program, so our instrumentation code knows
which registers and stack locations are available.

A side effect of this is that we have logic to show the control flow of
a program. Since we have this available, we can use this logic to draw
control flow graphs much like you would see in debuggers like Ida and
Ghidra. Perhaps someone has a use for this.