Added support for BPF-to-BPF functions in capture context
This commit implements BPF-to-BPF function call support, we do so by
passing the fp0 pointer to all functions. Before we used to just use
fp-256 to store fp0. Doing so caused a lot of stack memory usage which
made the verifier fail. This commit "solves" the issue by attempting to
use the stack directly adjecent the memory already used by the plain
program.

This requires us to know which parts of the stack are in use. To do so a
very limited program state analyser was implemented, inspired by how the
verifier in the kernel works. There is still some work to be done to
make it feature complete, but it works on a lot of programs already.