Updated ebpf loader library to latest and fixed BTF not found bug
The latest version of cilium/ebpf has made the btf package public
allowing us to access more BTF types which are needed for some features
we used to have in gobpfld.

With this bump, we also fixed a related issue which where BTF filenames
were not found and thus no source code was displayed.