Added local variables command
This commit adds the `locals`/`lv` command which lists all variables and
parameters in the current scope/function. When available the command
will show the current value of the var/param in memory.

Param/vars which are `not available` can be shown at some point in the
code, but are not in memory/registers at this moment. Params/vars marked
`inlined` are inlined variables which means they are removed from
compiled object altogether or are encoded in the instructions themselves

Fixes: #13